+++ zcash_primitives/src/sapling/note.rs
@@ -16,7 +16,7 @@ pub(super) mod nullifier;
 /// Before ZIP 212, the note commitment trapdoor `rcm` must be a scalar value.
 /// After ZIP 212, the note randomness `rseed` is a 32-byte sequence, used to derive
 /// both the note commitment trapdoor `rcm` and the ephemeral private key `esk`.
-#[derive(Copy, Clone, Debug)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub enum Rseed {
     BeforeZip212(jubjub::Fr),
     AfterZip212([u8; 32]),